added maxlengths to string and added missing mappings in dbcontext

+++ Synigo.OpenEducationApi/Model/V4/Models/CourseResult.cs
@@ -10,6 +10,6 @@ namespace Synigo.OpenEducationApi.Model.V4
         /// </summary>
         [Required]
         [JsonPropertyName("ects")]
-        public float Ects { get; set; }
+        public decimal Ects { get; set; }
     }
 }
